LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/alignment/flex-align-baseline-flex-004-expected.txt


PASS .target > * 1
PASS .target > * 2
FAIL .target > * 3 assert_equals:
<div data-offset-x="110"><span></span></div>
offsetLeft expected 110 but got 130
PASS .target > * 4
FAIL .target > * 5 assert_equals:
<div data-offset-x="105"><span></span></div>
offsetLeft expected 105 but got 40
PASS .target > * 6
FAIL .target > * 7 assert_equals:
<div data-offset-x="105"><span></span></div>
offsetLeft expected 105 but got 60
PASS .target > * 8
FAIL .target > * 9 assert_equals:
<div data-offset-x="45"><span></span></div>
offsetLeft expected 45 but got 65
PASS .target > * 10
PASS .target > * 11
PASS .target > * 12
FAIL .target > * 13 assert_equals:
<div data-offset-x="40"><span></span></div>
offsetLeft expected 40 but got 125
PASS .target > * 14
FAIL .target > * 15 assert_equals:
<div data-offset-x="40"><span></span></div>
offsetLeft expected 40 but got 105
PASS .target > * 16
